requirements.txt
streamlit
pandas
pyyaml
streamlit-authenticator
openpyxl
